strlen

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rgerdzhi <rgerdzhi@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 16:16:22 by rgerdzhi          #+#    #+#             */
-/*   Updated: 2024/06/25 16:49:00 by rgerdzhi         ###   ########.fr       */
+/*   Updated: 2024/06/25 18:56:26 by rgerdzhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (s[length] != '\0')
